+++ javascript/node/selenium-webdriver/phantomjs.js
@@ -104,8 +104,23 @@ var WEBDRIVER_TO_PHANTOMJS_LEVEL = (function() {
  * @param {webdriver.promise.ControlFlow=} opt_flow The control flow to use, or
  *     {@code null} to use the currently active flow.
  * @return {!webdriver.WebDriver} A new WebDriver instance.
+ * @deprecated Use {@link Driver}.
  */
 function createDriver(opt_capabilities, opt_flow) {
+  return new Driver(opt_capabilities, opt_flow);
+}
+
+
+/**
+ * Creates a new WebDriver client for PhantomJS.
+ *
+ * @param {webdriver.Capabilities=} opt_capabilities The desired capabilities.
+ * @param {webdriver.promise.ControlFlow=} opt_flow The control flow to use, or
+ *     {@code null} to use the currently active flow.
+ * @constructor
+ * @extends {webdriver.WebDriver}
+ */
+var Driver = function(opt_capabilities, opt_flow) {
   var capabilities = opt_capabilities || webdriver.Capabilities.phantomjs();
   var exe = findExecutable(capabilities.get(BINARY_PATH_CAPABILITY));
   var args = ['--webdriver-logfile=' + DEFAULT_LOG_FILE];
@@ -157,15 +172,22 @@ function createDriver(opt_capabilities, opt_flow) {
   var executor = executors.createExecutor(service.start());
   var driver = webdriver.WebDriver.createSession(
       executor, capabilities, opt_flow);
-  var boundQuit = driver.quit.bind(driver);
-  driver.quit = function() {
+
+  webdriver.WebDriver.call(
+      this, driver.getSession(), executor, driver.controlFlow());
+
+  var boundQuit = this.quit.bind(this);
+
+  /** @override */
+  this.quit = function() {
     return boundQuit().thenFinally(service.kill.bind(service));
   };
   return driver;
-}
+};
+util.inherits(Driver, webdriver.WebDriver);
 
 
 // PUBLIC API
 
-
+exports.Driver = Driver;
 exports.createDriver = createDriver;